CFTimeSchemes

Documentation for CFTimeSchemes.

CFTimeSchemes.IVPSolver

CFTimeSchemes.RungeKutta4

CFTimeSchemes.advance!

CFTimeSchemes.advance!

CFTimeSchemes.model_dstate

CFTimeSchemes.scratch_space

CFTimeSchemes.tendencies!

CFTimeSchemes.time_stages

CFTimeSchemes.update!

CFTimeSchemes.IVPSolver — Type

solver = IVPSolver(model, scheme, dt, u0, mutating=false)
solver = IVPSolver(model, scheme, dt)

Use solver with advance.

By default, solver is non-mutating, which is known to work with Zygote but allocates. If u0 is provided and mutating=true, advance should not allocate and have better performance. The type and shape of u0 (but not the values) are used to allocate scratch spaces, see scratch_space. This allows the non-mutating mode to work with ForwardDiff.

CFTimeSchemes.RungeKutta4 — Type

scheme = RungeKutta4(model)

4th-order Runge Kutta scheme for model. Pass scheme to IVPSolver.

CFTimeSchemes.advance! — Function

future, t = advance!(future, scheme, present, t, dt, scratch)
future, t = advance!(void, scheme, present, t, dt, void)

Integrate in time by one time step, respectively mutating (non-allocating) and non-mutating

CFTimeSchemes.advance! — Method

future, t = advance!(future, solver, present, t, N)
future, t = advance!(void, solver, present, t, N)

CFTimeSchemes.model_dstate — Function

k = model_dstate(model, u)

Returns an object in which we can store tendencies ; u0 is passed to handle ForwardDiff.Dual

CFTimeSchemes.scratch_space — Function

space = scratch_space(model, u0)
space = scratch_space(scheme, u0)

scratch space used to compute tendencies or keep sub-stages (RK); u0 is passed to handle ForwardDiff.Dual

CFTimeSchemes.tendencies! — Function

dstate = tendencies!(dstate, model, u, scratch, t) # Mutating
dstate = tendencies!(void, model, u, void, t) # Non-mutatingœ

Computes tendencies. Pass void to arguments dstate and scratch for non-mutating variant.

CFTimeSchemes.time_stages — Function

stages = time_stages(scheme, model, u0)

returns an object storing tendencies at sub-stages (RK). u0 is passed to handle ForwardDiff.Dual

CFTimeSchemes.update! — Method

new = update!(new, new, increment, factor)
new = update!(new, old, increment, factor)
new = update!(void, old, increment, factor)

Respectively equivalent to: @. new += factorincrement @. new = old + factorincrement new = @. old + factor*increment Operates recursively on nested tuples / named tuples